+++ paq8pxd.cpp
@@ -138,6 +138,16 @@ inline int max(int a, int b) {return a<b?b:a;}
 
 #include "predictors/predictors.hpp"
 #include "predictors/predictor.hpp"
+#include "predictors/predictordec.hpp"
+#include "predictors/predictorjpeg.hpp"
+#include "predictors/predictorexe.hpp"
+#include "predictors/predictorimg4.hpp"
+#include "predictors/predictorimg8.hpp"
+#include "predictors/predictorimg24.hpp"
+#include "predictors/predictortext.hpp"
+#include "predictors/predictorimg1.hpp"
+#include "predictors/predictoraudio.hpp"
+
 /////////////////////// Global context /////////////////////////
 U8 level=DEFAULT_OPTION;  // Compression level 0 to 15
 bool slow=false; //-x
@@ -233,851 +243,12 @@ size_t getPeakMemory(){
 
 
 
-// DECAlpha predicor
-class PredictorDEC: public Predictors {
-  int pr;
-  int pr0;
-  int order;
-  int ismatch;
-  Mixer *m; 
-  struct {
-      APM APMs[1];
-    } DEC;
-  eSSE sse;
-  const U8 activeModels[15] = { 
-   M_RECORD,
-   M_MATCH ,
-   M_MATCH1, 
-   M_DISTANCE,  
-   M_INDIRECT, 
-   M_DMC, 
-   M_NEST, 
-   M_NORMAL,  
-   M_TEXT, 
-   M_WORD ,
-   M_DEC,  
-   M_SPARSEMATCH, 
-   M_SPARSE_Y,
-   M_CHART,M_LSTM    };
-public:
-  PredictorDEC();
-  int p()  const {/*assert(pr>=0 && pr<4096);*/ return pr;} 
-  void update() ;
-   ~PredictorDEC(){
-  }
-};
 
-PredictorDEC::PredictorDEC(): pr(16384),pr0(pr),order(0),ismatch(0),
-  DEC{
-    { /*APM:*/ { 25*26,20} }
-  },
-  sse(x){
-   loadModels(activeModels,15);
-   // add extra 
-   mixerInputs+=1;
-   mixerNets+= (8+1024)+       256+   256+   256+    256+    1536+64;
-   mixerNetsCount+=7;
-   sse.p(pr);
-   // create mixer
-   m=new Mixer(mixerInputs,  mixerNets,x, mixerNetsCount,0,3,2);
-}
 
-void PredictorDEC::update()  {
-    if (x.bpos==0) {
-        int b1=x.buf(1);
-        int i=WRT_mpw[b1>>4];
-        x.w4=x.w4*4+i;
-        if (x.b2==3) i=2;
-        x.w5=x.w5*4+i;
-        x.b3=x.b2;
-        x.b2=b1;   
-        x.x4=x.x4*256+b1,x.x5=(x.x5<<8)+b1;
-        if(b1=='.' || b1=='!' || b1=='?' || b1=='/'|| b1==')'|| b1=='}') {
-            x.w5=(x.w5<<8)|0x3ff,x.f4=(x.f4&0xfffffff0)+2,x.x5=(x.x5<<8)+b1,x.x4=x.x4*256+b1;
-            if(b1!='!') x.w4|=12, x.tt=(x.tt&0xfffffff8)+1,x.b3='.';
-        }
-        if (b1==32) --b1;
-        x.tt=x.tt*8+WRT_mtt[b1];
-        x.f4=x.f4*16+(b1>>4);
-    }
 
-    m->update();
-    m->add(256);
-    ismatch=models[M_MATCH]->p(*m);
-    models[M_MATCH1]->p(*m);
-    if (x.bpos==0)x.count++;
-    models[M_SPARSEMATCH]->p(*m);
-    order=models[M_NORMAL]->p(*m);
-    models[M_RECORD]->p(*m,4 );
-    models[M_TEXT]->p(*m );
-    models[M_WORD]->p(*m );
-    models[M_NEST]->p(*m );
-    models[M_SPARSE_Y]->p(*m,ismatch,order);
-    models[M_DISTANCE]->p(*m);
-    models[M_INDIRECT]->p(*m);
-    models[M_DMC]->p(*m);
-    models[M_DEC]->p(*m);
-    m->set(order << 3U | x.bpos, 64);
-    if (slow==true) models[M_CHART]->p(*m);
-    if (slow==true) models[M_LSTM]->p(*m);
-    U32 c1=x.buf(1), c2=x.buf(2), c3=x.buf(3), c;
-    m->set(8+ c1 + (x.bpos>5)*256 + ( ((x.c0&((1<<x.bpos)-1))==0) || (x.c0==((2<<x.bpos)-1)) )*512, 8+1024);
-    m->set(x.c0, 256);
-    m->set(order+8*(x.c4>>6&3)+32*(x.bpos==0)+64*(c1==c2)| (1)<<7, 256);  
-    m->set(c2, 256);
-    m->set(c3, 256);
-    if (x.bpos) {
-      c=x.c0<<(8-x.bpos); if (x.bpos==1)c+=c3/2;
-      c=(min(x.bpos,5))*256+c1/32+8*(c2/32)+(c&192);
-    }
-    else c=c3/128+(x.c4>>31)*2+4*(c2/64)+(c1&240); 
-    m->set(c, 1536);
-    pr0=m->p(0,1); //0,1
-    int const limit = 0x3FFu >> ((x.blpos < 0xFFFu) * 4);
-    pr = DEC.APMs[0].p(pr0, (x.DEC.state * 26u) + x.DEC.bcount, x.y,limit);
-    pr = (pr * 4 + pr0 * 2 + 3) / 6;
-    pr=(4096-pr)*(32768/4096);
-    if(pr<1) pr=1;
-    if(pr>32767) pr=32767;
-}
-// Jpeg predicor
-class PredictorJPEG: public Predictors {
-  int pr;
-  Mixer *m;
-  struct {
-      APM APMs[1];
-    } Jpeg;
-  bool Bypass; 
-  const U8 activeModels[5] = { 
-   M_JPEG,
-   M_MATCH ,
-   M_MATCH1, 
-   M_NORMAL, 
-   M_LSTM    };
-public:
-  int p()  const {/*assert(pr>=0 && pr<4096);*/ return pr;} 
-  ~PredictorJPEG(){   //printf("\n JPEG Count of skipped bytes %d\n",x.count);
-      
- }
-PredictorJPEG(): pr(16384), 
-  Jpeg{
-    { /*APM:*/ { 0x2000} }
-  },
-  Bypass(false){
-   loadModels(activeModels,5);
-   // add extra 
-   mixerInputs+=3+1+1-3;
-   mixerNets+=        (8+1024)+       256+       256+       256+       256+       1536;
-     
-   mixerNetsCount+=6;
-   // create mixer
-   m=new Mixer(mixerInputs,  mixerNets,x, mixerNetsCount,0,3,2); //set  error update rate to 1.5 (3/2), default is 7/1
-}
 
-void update()  {
-    pr=(32768-pr)/(32768/4096);
-    if(pr<1) pr=1;
-    if(pr>4095) pr=4095;
-    x.Misses+=x.Misses+((pr>>10)!=x.y);
-    m->update();
-    m->add(256);
-    Bypass=false;
-    int ismatch=models[M_MATCH]->p(*m);
-    models[M_MATCH1]->p(*m);
-    if (slow==true) models[M_LSTM]->p(*m);
-    if (slow==false && (x.Match.length>0xFF || x.Match.bypass)) {//256b
-        x.Match.bypass =   Bypass =    true;
-        pr= x.Match.bypassprediction;
-        pr=(4096-pr)*(32768/4096);
-        if(pr<1) pr=1;
-        if(pr>32767) pr=32767;
-        models[M_JPEG]->p(*m,1);//we found long repeating match. update, do not predict. artificial images, same partial content
-        m->reset(); 
-        return;
-    }
-    if (models[M_JPEG]->p(*m)) {
-        if (slow==true) models[M_NORMAL]->p(*m);
-        pr=m->p(1,0);
-    }
-    else{
-        int order =models[M_NORMAL]->p(*m);
-        U32 c1=x.buf(1), c2=x.buf(2), c3=x.buf(3), c;
-        m->set(8+ c1 + (x.bpos>5)*256 + ( ((x.c0&((1<<x.bpos)-1))==0) || (x.c0==((2<<x.bpos)-1)) )*512, 8+1024);
-        m->set(x.c0, 256);
-        m->set(order | ((x.c4>>6)&3)<<3 | (x.bpos==0)<<5 | (c1==c2)<<6 | (1)<<7, 256);
-        m->set(c2, 256);
-        m->set(c3, 256);
-        U8 d=x.c0<<(8-x.bpos);
-        if (x.bpos) {
-            c=d; if (x.bpos==1)c+=c3/2;
-            c=(min(x.bpos,5))*256+c1/32+8*(c2/32)+(c&192);
-        }
-        else c=c3/128+(x.c4>>31)*2+4*(c2/64)+(c1&240); 
-        m->set(c, 1536);
-        pr=m->p(1,1);
-    }
-    U32 pr0 = Jpeg.APMs[0].p(pr , x.JPEG.state,x.y, 0x3FF);
-    pr = (pr + pr0 + 1) / 2;
-    pr=(4096-pr)*(32768/4096);
-    if(pr<1) pr=1;
-    if(pr>32767) pr=32767;
-}
-};
 
-// x86/64 predicor
-class PredictorEXE: public Predictors {
-  int pr;  // next prediction
-  int order;
-  Mixer *m;
-  struct {
-      APM APMs[3];
-    } x86_64;
-  U32 count;
-  eSSE sse;
-  const U8 activeModels[17] = { 
-   M_RECORD,
-   M_MATCH ,
-   M_MATCH1, 
-   M_DISTANCE, 
-   M_EXE, 
-   M_INDIRECT, 
-   M_DMC, 
-   M_NEST, 
-   M_NORMAL, 
-   M_XML, 
-   M_TEXT, 
-   M_WORD ,
-   M_SPARSEMATCH, 
-   M_SPARSE_Y,
-   M_PPM,M_CHART,M_LSTM };
-public:
-  PredictorEXE();
-  int p()  const {/*assert(pr>=0 && pr<4096);*/ return pr;} 
-  void update() ;
-    ~PredictorEXE(){
-    }
-};
 
-PredictorEXE::PredictorEXE(): pr(16384),order(0),
-  x86_64{
-    { /*APM:*/ {0x800,20}, {0x10000,16}, {0x10000,16} }
-  },
-  count(0), sse(x) {
-  loadModels(activeModels,17);
-   // add extra 
-   mixerInputs+=1;
-   mixerNets+=( 8+1024)+     256+     256+     256+     256+     256+     1536;
-   mixerNetsCount+=7;
-   sse.p(pr);
-   // create mixer
-   m=new Mixer(mixerInputs,  mixerNets,x, mixerNetsCount);
-}
-
-void PredictorEXE::update()  {
-    pr=(32768-pr)/(32768/4096);
-    if(pr<1) pr=1;
-    if(pr>4095) pr=4095;
-    x.Misses+=x.Misses+((pr>>11)!=x.y);
-    if (x.bpos==0) {
-        int b1=x.buf(1);
-        int i=WRT_mpw[b1>>4];
-        x.w4=x.w4*4+i;
-        if (x.b2==3) i=2;
-        x.w5=x.w5*4+i;
-        x.b3=x.b2;
-        x.b2=b1;   
-        x.x4=x.x4*256+b1,x.x5=(x.x5<<8)+b1;
-        if(b1=='.' || b1=='!' || b1=='?' || b1=='/'|| b1==')'|| b1=='}') {
-            x.w5=(x.w5<<8)|0x3ff,x.f4=(x.f4&0xfffffff0)+2,x.x5=(x.x5<<8)+b1,x.x4=x.x4*256+b1;
-            if(b1!='!') x.w4|=12, x.tt=(x.tt&0xfffffff8)+1,x.b3='.';
-        }
-        if (b1==32) --b1;
-        x.tt=x.tt*8+WRT_mtt[b1];
-        x.f4=x.f4*16+(b1>>4);
-    }
-    m->update();
-    m->add(256);
-     
-    int ismatch=models[M_MATCH]->p(*m);
-    models[M_MATCH1]->p(*m);
-    if (x.bpos==0)x.count++;
-    models[M_SPARSEMATCH]->p(*m);
-    order=models[M_NORMAL]->p(*m);
-    int rec=0;
-    rec=models[M_RECORD]->p(*m);
-    models[M_WORD]->p(*m,order);
-    models[M_NEST]->p(*m);
-    models[M_SPARSE_Y]->p(*m,ismatch,order);
-    models[M_DISTANCE]->p(*m);
-    models[M_INDIRECT]->p(*m);
-    models[M_DMC]->p(*m);
-    models[M_EXE] ->p(*m,1);
-    if (slow==false) models[M_XML]->p(*m);
-    models[M_TEXT]->p(*m);
-    if (slow==true) models[M_PPM]->p(*m); 
-    if (slow==true) models[M_CHART]->p(*m);
-    if (slow==true) models[M_LSTM]->p(*m);
-    U32 c1=x.buf(1), c2=x.buf(2), c3=x.buf(3), c;
-    m->set(8+ c1 + (x.bpos>5)*256 + ( ((x.c0&((1<<x.bpos)-1))==0) || (x.c0==((2<<x.bpos)-1)) )*512, 8+1024);
-    m->set(x.c0, 256);
-    m->set(c2, 256);
-    U8 d=x.c0<<(8-x.bpos);
-    m->set(order+8*(x.c4>>6&3)+32*(x.bpos==0)+64*(c1==c2)+1*128, 256);
-    m->set(c3, 256);
-    m->set(ismatch, 256);
-    if (x.bpos) {
-        c=d; if (x.bpos==1)c+=c3/2;
-        c=(min(x.bpos,5))*256+c1/32+8*(c2/32)+(c&192);
-    }
-    else c=c3/128+(x.c4>>31)*2+4*(c2/64)+(c1&240); 
-    m->set(c, 1536);
-    int pr0=m->p(1,1);
-    int const limit = 0x3FFu >> (static_cast<int>(x.blpos < 0xFFFu) * 4);
-    pr = x86_64.APMs[0].p(pr0, (x.x86_64.state << 3u) | x.bpos,x.y, limit);
-    int  pr1 = x86_64.APMs[1].p(pr0, (x.c0 << 8u) | x.x86_64.state,x.y, limit);
-    int  pr2 = x86_64.APMs[2].p((pr1+pr+1)/2, finalize64(hash(x.c4 & 0xFFu, x.bpos, x.Misses & 0x1u, x.x86_64.state >> 3), 16),x.y, limit);
-    pr = (pr + pr0 + pr1 + pr2 + 2) >> 2;
-    sse.update();
-    pr = sse.p(pr);
-}
-
-// 4-bit image predicor
-class PredictorIMG4: public Predictors {
-  int pr;
-  Mixer *m;
-  StateMap StateMaps[2];
-  struct {
-      APM APMs[4];
-      APM1 APM1s[2];
-  } Image;
-  eSSE sse;
-  const U8 activeModels[4] = { 
-   M_MATCH ,
-   M_MATCH1, 
-   M_IM4,
-   M_LSTM    };
-public:
-  int p()  const {/*assert(pr>=0 && pr<4096);*/ return pr;} 
-   ~PredictorIMG4(){ }
-
-PredictorIMG4(): pr(16384), StateMaps{ 256, 256*256}, Image
-     {{0x1000, 0x8000, 0x8000, 0x8000},  {{0x10000,x}, {0x10000,x}}}, sse(x) {
-   loadModels(activeModels,4);  
-   // add extra 
-   mixerInputs+=1+2;
-   mixerNets+=0;
-   mixerNetsCount+=0;
-   sse.p(pr);
-   // create mixer
-   m=new Mixer(mixerInputs,  mixerNets,x, mixerNetsCount);
-}
-
-void update()  {
-  pr=(32768-pr)/(32768/4096);
-  if(pr<1) pr=1;
-  if(pr>4095) pr=4095;
-  x.Misses+=x.Misses+((pr>>11)!=x.y);
-  m->update();
-  m->add(256);
-  models[M_MATCH]->p(*m);
-  models[M_MATCH1]->p(*m);
-  models[M_IM4]->p(*m,x.finfo);
-  if (slow==true) models[M_LSTM]->p(*m);
-  m->add((stretch(StateMaps[0].p(x.c0,x.y))+1)>>1);
-  m->add((stretch(StateMaps[1].p(x.c0|(x.buf(1)<<8),x.y))+1)>>1);
-  int pr0=m->p();
-  int pr1, pr2, pr3;
-  int limit=0x3FF>>((x.blpos<0xFFF)*4);
-  pr  = Image.APMs[0].p(pr0, (x.c0<<4)|(x.Misses&0xF), x.y, limit);
-  pr1 = Image.APMs[1].p(pr0, hash(x.c0, x.Image.pixels.W, x.Image.pixels.N)&0x7FFF,x.y, limit);
-  pr2 = Image.APMs[2].p(pr0, hash(x.c0, x.Image.pixels.N, x.Image.pixels.NN)&0x7FFF,x.y, limit);
-  pr3 = Image.APMs[3].p(pr0, hash(x.c0, x.Image.pixels.W, x.Image.pixels.WW)&0x7FFF,x.y, limit);
-  pr0 = (pr0+pr1+pr2+pr3+2)>>2;
-      
-  pr1 = Image.APM1s[0].p(pr0, hash(x.c0, x.Match.byte, x.Image.pixels.N)&0xFFFF, 5);
-  pr2 = Image.APM1s[1].p(pr, hash(x.c0, x.Image.pixels.W, x.Image.pixels.N)&0xFFFF, 6);
-  pr = (pr*2+pr1+pr2+2)>>2;
-  pr = (pr+pr0+1)>>1;   
-  sse.update();
-  pr = sse.p(pr);
-}
-};
-
-// 8-bit image predicor
-class PredictorIMG8: public Predictors {
-  int pr;
-  Mixer *m;
-  struct {
-    struct {
-      APM APMs[4];
-      APM1 APM1s[2];
-    } Palette;
-      struct {
-      APM APMs[3];
-    } Gray;
-  } Image;
-  StateMap StateMaps[2];
-  eSSE sse;
-  const U8 activeModels[4] = { 
-   M_MATCH ,
-   M_MATCH1, 
-   M_IM8,
-   //M_NORMAL,
-   M_LSTM    };
-public:
-  int p()  const {/*assert(pr>=0 && pr<4096);*/ return pr;} 
-   ~PredictorIMG8(){ }
-
-PredictorIMG8(): pr(16384),  Image{
-     {{0x1000, 0x10000, 0x10000, 0x10000},  {{0x10000,x}, {0x10000,x}}},
-     {0x1000, 0x10000, 0x10000} } ,StateMaps{ 256, 256*256}, sse(x){
-  loadModels(activeModels,4);  
-   // add extra 
-   mixerInputs+=1+2;
-   sse.p(pr);x.count=0x1ffff;
-   // create mixer
-   m=new Mixer(mixerInputs,  mixerNets,x, mixerNetsCount,0,3,2);
-}
-
-void update()  {
-  pr=(32768-pr)/(32768/4096);
-  if(pr<1) pr=1;
-  if(pr>4095) pr=4095;
-  x.Misses+=x.Misses+((pr>>11)!=x.y);
-  m->update();
-  m->add(256);
-  if (x.bpos==0)x.count++;
-  models[M_MATCH]->p(*m);
-  models[M_MATCH1]->p(*m);
-  //if (slow==true) models[M_NORMAL]->p(*m);
-  models[M_IM8]->p(*m,x.finfo);
-  if (slow==true) models[M_LSTM]->p(*m);
-  m->add((stretch(StateMaps[0].p(x.c0,x.y))+1)>>1);
-  m->add((stretch(StateMaps[1].p(x.c0|(x.buf(1)<<8),x.y))+1)>>1);
-  
-   if(x.filetype== IMAGE8GRAY)  {
-      int pr0=m->p(0,1);
-      int pr1, pr2, pr3;
-      int limit=0x3FF>>((x.blpos<0xFFF)*4);
-      pr  = Image.Gray.APMs[0].p(pr0, (x.c0<<4)|(x.Misses&0xF), x.y,limit);
-      pr1 = Image.Gray.APMs[1].p(pr, (x.c0<<8)|x.Image.ctx, x.y,limit);
-      pr2 = Image.Gray.APMs[2].p(pr0, x.bpos|(x.Image.ctx&0xF8)|(x.Match.byte<<8),x.y, limit);
-      pr0 = (2*pr0+pr1+pr2+2)>>2;
-      pr = (pr+pr0+1)>>1; 
-      }
-  else {
-       int pr0=m->p(1,1);
-      int pr1, pr2, pr3;
-      int limit=0x3FF>>((x.blpos<0xFFF)*4);
-      pr  = Image.Palette.APMs[0].p(pr0, (x.c0<<4)|(x.Misses&0xF), x.y, limit);
-      pr1 = Image.Palette.APMs[1].p(pr0, hash(x.c0, x.Image.pixels.W, x.Image.pixels.N)&0xFFFF,x.y, limit);
-      pr2 = Image.Palette.APMs[2].p(pr0, hash(x.c0, x.Image.pixels.N, x.Image.pixels.NN)&0xFFFF,x.y, limit);
-      pr3 = Image.Palette.APMs[3].p(pr0, hash(x.c0, x.Image.pixels.W, x.Image.pixels.WW)&0xFFFF,x.y, limit);
-      pr0 = (pr0+pr1+pr2+pr3+2)>>2;
-      
-      pr1 = Image.Palette.APM1s[0].p(pr0, hash(x.c0, x.Match.byte, x.Image.pixels.N)&0xFFFF, 5);
-      pr2 = Image.Palette.APM1s[1].p(pr, hash(x.c0, x.Image.pixels.W, x.Image.pixels.N)&0xFFFF, 6);
-      pr = (pr*2+pr1+pr2+2)>>2;
-      pr = (pr+pr0+1)>>1;   
-  }
-  sse.update();
-  pr = sse.p(pr);
-}
-};
-
-// 24/32-bit image predicor
-class PredictorIMG24: public Predictors {
-  int pr;  // next prediction
-  Mixer *m;
-  struct {
-    APM APMs[1];
-    APM1 APM1s[2];
-  } Image;
-  StateMap StateMaps[2];
-  eSSE sse;
-  const U8 activeModels[4] = { 
-   M_MATCH ,
-   M_MATCH1, 
-   M_IM24,
-   //M_NORMAL,
-   M_LSTM    };
-public:
-  int p()  const {/*assert(pr>=0 && pr<4096);*/ return pr;} 
-   ~PredictorIMG24(){ }
-
-PredictorIMG24(): pr(16384),Image{ {0x1000/*, 0x10000, 0x10000, 0x10000*/}, {{0x10000,x}, {0x10000,x}} },
-                  StateMaps{ 256, 256*256}, sse(x){
-  loadModels(activeModels,4);   
-   // add extra 
-   mixerInputs+=1+2;
-   mixerNets+=  8192;
-   mixerNetsCount+=1;
-   sse.p(pr);
-   // create mixer
-   m=new Mixer(mixerInputs,  mixerNets,x, mixerNetsCount,0,3,2);
-}
-
-void update()  {
-  pr=(32768-pr)/(32768/4096);
-  if(pr<1) pr=1;
-  if(pr>4095) pr=4095;
-  
-  x.Misses+=x.Misses+((pr>>11)!=x.y);
-  m->update();
-  m->add(256);
-  if (x.bpos==0)x.count++;
-  models[M_MATCH]->p(*m);
-  models[M_MATCH1]->p(*m);
-  //if (slow==true) models[M_NORMAL]->p(*m);
-  if (slow==true) models[M_LSTM]->p(*m);
-  models[M_IM24]->p(*m,x.finfo);
-  m->add((stretch(StateMaps[0].p(x.c0,x.y))+1)>>1);
-  m->add((stretch(StateMaps[1].p(x.c0|(x.buf(1)<<8),x.y))+1)>>1);
-  m->set(x.Image.ctx&0x1FFF,8192);
-  int pr1, pr2, pr3;
-  int pr0=x.filetype==IMAGE24? m->p(1,1): m->p();
-  int limit=0x3FF>>((x.blpos<0xFFF)*4);
- // pr=pr0;
-  pr  = Image.APMs[0].p(pr0, (x.c0<<4)|(x.Misses&0xF), x.y,limit);
- /* pr1 = Image.APMs[1].p(pr0, hash(x.c0, x.Image.pixels.W, x.Image.pixels.WW)&0xFFFF,  x.y,limit);
-  pr2 = Image.APMs[2].p(pr0, hash(x.c0, x.Image.pixels.N, x.Image.pixels.NN)&0xFFFF, x.y, limit);
-  pr3 = Image.APMs[3].p(pr0, (x.c0<<8)|x.Image.ctx, x.y, limit);
-  pr0 = (pr0+pr1+pr2+pr3+2)>>2;
-*/
-  pr1 = Image.APM1s[0].p(pr, hash(x.c0, x.Image.pixels.W, x.buf(1)-x.Image.pixels.Wp1, x.Image.plane)&0xFFFF);
-  pr2 = Image.APM1s[1].p(pr, hash(x.c0, x.Image.pixels.N, x.buf(1)-x.Image.pixels.Np1, x.Image.plane)&0xFFFF);
-  pr=(pr*2+pr1*3+pr2*3+4)>>3;
-  pr = (pr+pr0+1)>>1;
-  sse.update();
-  pr = sse.p(pr);
-}
-};
-
-// Text predicor
-class PredictorTXTWRT: public Predictors {
-  int pr;  // next prediction
-  int pr0;
-  int order;
-  int rlen;
-  int ismatch;
-  Mixer *m;
-  struct {
-    APM APMs[4];
-    APM1 APM1s[3];
-  } Text;
-  U32 count;
-  U32 blenght;
-  StateMap StateMaps[1];
-  wrtDecoder wr;
-  eSSE sse;
-  int decodedTextLen,lasttag;
-  int counttags,lState;
-  const U8 activeModels[15] = { 
-   M_RECORD,
-   M_MATCH ,
-   M_MATCH1,  
-   M_INDIRECT, 
-   M_DMC, 
-   M_NEST, 
-   M_NORMAL, 
-   M_XML, 
-   M_TEXT, 
-   M_WORD , 
-   M_SPARSEMATCH, 
-   M_SPARSE,
-   M_PPM,M_CHART,M_LSTM    };
-public:
-  PredictorTXTWRT();
-  int p()  const {/*assert(pr>=0 && pr<4096);*/ return pr;} 
-  void wrt();
-  void update() ;
-   ~PredictorTXTWRT(){
-      // printf("\n TXTWRT Count of skipped bytes %d\n",count/8);
-  }
-};
-
-PredictorTXTWRT::PredictorTXTWRT(): pr(16384),pr0(pr),order(0),rlen(0),ismatch(0),
-Text{ {0x10000, 0x10000, 0x10000, 0x10000}, {{0x10000,x}, {0x10000,x}, {0x10000,x}} },
-count(0),blenght(1024*4),StateMaps{  ( 0x7FFFFF+1)<<2},sse(x),
-decodedTextLen(0),lasttag(0),counttags(0),lState(0){
-   loadModels(activeModels,15);  
-   // add extra 
-   mixerInputs+=1+1;
-   mixerNets+= 4096+ 64+ 256+ 256*8+ 256*8+ 256*8+ 1536+ 2048;
-   mixerNetsCount+=8;
-   sse.p(pr); // must
-   // create mixer
-   m=new Mixer(mixerInputs,  mixerNets,x, mixerNetsCount,0,3,2);
-}
-void PredictorTXTWRT::wrt(){
-        U8 c1=x.c4;
-        if (x.wrtLoaded==true) {
-           if (c1=='<' && x.ishtml==true ) x.istex=false; // open
-        }
-    // load wrt dictionary from file
-        // detect header//'!Pq8'
-        if (x.c4==0x21507138 && x.blpos<16){ 
-          x.wrtpos=x.blpos;
-        }
-        if (x.wrtpos>0)  { 
-          if (x.wrtpos+4+4==x.blpos){ 
-            x.wrtfile=(U64(bswap(x.c4))<<32)| bswap(x.c8);
-          }
-          //load size
-          if (x.wrtpos+4+4+3==x.blpos) { 
-          x.wrtsize=(x.c4&0xffffff);
-          if (x.wrtsize<4) x.wrtpos=0;
-          }
-          //load count
-          if (x.wrtpos+4+4+3+3==x.blpos) { 
-            x.wrtcount=(x.c4&0xffffff);
-            x.wrtcount=(x.wrtcount>>16)+((x.wrtcount&0xff)<<16)+(x.wrtcount&0xff00);
-            printf("WRT dict count %d words.\n",x.wrtcount);
-            x.wwords=x.wrtcount;
-            if (x.wrtcount==0) x.wrtpos=x.wrtsize=0;
-          }
-          if (x.wrtsize>0 && x.wrtpos+4+4+3+3+x.wrtsize+3==x.blpos && x.wrtdata==0) { 
-            if (x.buf(6)==5) {
-              x.wrtdata=x.blpos+5;
-            }
-            else x.wrtLoaded==false,x.wrtpos=x.wrtdata=x.wrtsize=0;
-          }
-          if (x.wrtdata>0 && x.blpos>=x.wrtdata &&x.wrtLoaded==false) { 
-            wr.WRT_start_decoding(x.wrtcount,x.wrtsize,&x.buf[x.buf.pos-5-3-x.wrtsize]);
-            x.wrtLoaded=x.dictonline=true;
-            printf("WRT dict online.\n");
-          }
-        }
-        if (x.wrtLoaded==true) {
-           // Returned wrtstatus values
-           // codeword size   1-3
-           // D_REGULAR_CHAR    0
-           // D_WORD_DECODED   -1     
-           // D_FIRSTUPPER     -2
-           // D_UPPERWORD      -3
-           // D_ESCAPE         -4
-           // D_UTFUPPER       -5
-          x.wrtstatus=wr.WRT_decode(x.buf(1),&x.wrtText[0],&x.wrtTextSize);
-          x.wstat=  (x.wrtstatus==-1)||  (x.wrtstatus>0);
-          x.wdecoded=false;
-          if (x.wrtstatus==-1){
-            x.wrtbytesize++;
-            x.wdecoded=true;
-            // print decoded word
-            // for (int p=0;p<x.wrtTextSize;p++)printf("%c",x.wrtText[p]);
-            models[M_TEXT]->setword(&x.wrtText[0],x.wrtTextSize);
-            x.wlen=x.wrtbytesize;
-             int y=0;
-            if (x.wrtTextSize<5 && x.wrtText[0]>'z') y=utf8_check(&x.wrtText[0]); 
-            if (y==x.wrtbytesize){
-                x.utf8l=y;
-                x.wrtc4=(x.wrtc4<<8)|x.wrtText[0];
-                x.bufn[x.bufn.pos++]=x.wrtc4&0xff;
-                x.bufn.pos=x.bufn.pos&x.bufn.poswr; //wrap
-            }else {
-            x.utf8l=0;
-            for (int p=0;p<x.wrtTextSize;p++){
-                U8 wc=x.wrtText[p];
-                x.wrtc4=(x.wrtc4<<8)|wc;
-                x.bufn[x.bufn.pos++]=wc;
-                x.bufn.pos=x.bufn.pos&x.bufn.poswr; //wrap
-            }
-            }
-            decodedTextLen=x.wrtTextSize;
-            x.wrtbytesize=0;
-            x.wrtTextSize=0;
-          }else if (x.wrtstatus>0){
-            x.wrtbytesize++;
-          }
-          // if line starts with <someword>
-          if ((x.buf(decodedTextLen)=='<' || x.bufn(decodedTextLen+1)=='<' /*|| (x.bufn(decodedTextLen+1)=='<' && x.bufn(decodedTextLen+2)=='&')*/) && c1=='>'){
-              counttags++;lasttag=x.blpos;
-          }
-          if ( c1 >'@' && lState==0)  lState=1;
-          if ((c1=='<' && lState==0)) lState=2;
-          if ((c1==10 || c1==5) && counttags && lState==2) {
-              x.ishtml=true;
-              counttags=0;
-          } else if (c1==10 || c1==5){ 
-               counttags=lState=0;
-               if ((x.blpos-lasttag)>256*4 && x.ishtml==true) x.ishtml=false;// standard break 1k
-          }
-        }
-
-        if (x.wrtstatus==0 && x.wrtLoaded==true){
-          if (c1==5) c1=10;
-          x.wrtc4=(x.wrtc4<<8)|c1;// printf("%c",c1);
-          x.bufn[x.bufn.pos++]=c1;
-          x.bufn.pos=x.bufn.pos&x.bufn.poswr; //wrap
-        }
-        if (c1=='>') x.istex=true;  //close
-        //if (x.c4==0x10103d3d) x.ishtml=false,x.istex=true;  // w - fast break
-}
-void PredictorTXTWRT::update() {
-    if (x.bpos==0) {
-        int b1=x.buf(1);
-        int i=WRT_mpw[b1>>4];
-        x.w4=x.w4*4+i;
-        if (x.b2==3) i=2;
-        x.w5=x.w5*4+i;
-        x.b4=x.b3;
-        x.b3=x.b2;
-        x.b2=b1;   
-        x.x4=x.x4*256+b1,x.x5=(x.x5<<8)+b1;
-        if(b1=='.' || b1=='!' || b1=='?' || b1=='/'|| b1==')'|| b1=='}') {
-            x.w5=(x.w5<<8)|0x3ff,x.f4=(x.f4&0xfffffff0)+2,x.x5=(x.x5<<8)+b1,x.x4=x.x4*256+b1;
-            if(b1!='!') x.w4|=12;//, x.tt=(x.tt&0xfffffff8)+1,x.b3='.';
-        }
-        
-        x.tt=x.tt*8+WRT_mtt[b1];
-        if (b1==32) --b1;
-        x.f4=x.f4*16+(b1>>4);
-        wrt();
-        int d=WRT_wrd1[b1]; 
-        x.bc4=x.bc4<<2|(d/64); 
-    }
-    pr=(32768-pr)/(32768/4096);
-    if(pr<1) pr=1;
-    if(pr>4095) pr=4095;
-    x.Misses+=x.Misses+((pr>>11)!=x.y);
-    m->update();
-    m->add(256);
-    ismatch=models[M_MATCH]->p(*m);
-    models[M_MATCH1]->p(*m);
-    if (x.bpos==0)x.count++;
-    models[M_SPARSEMATCH]->p(*m);
-    order=models[M_NORMAL]->p(*m);
-
-        int state=models[M_XML]->p(*m);
-        m->set(models[M_WORD]->p(*m,state&7)&0xfff, 4096);    
-        models[M_SPARSE]->p(*m,ismatch,order);
-        models[M_NEST]->p(*m);
-        models[M_INDIRECT]->p(*m);
-        models[M_DMC]->p(*m);
-        rlen=models[M_RECORD]->p(*m);
-        models[M_TEXT]->p(*m,(state&7));
-        if (slow==true) models[M_PPM]->p(*m);
-        if (slow==true) models[M_CHART]->p(*m);
-        if (slow==true) models[M_LSTM]->p(*m);
-        int dd=(pr>>9)>0+(pr>>9)>14;//+pr>7;
-        dd=(dd<<7)|(x.bc4&127);
-        dd=(dd<<8)|(x.buf(1));
-        dd=(dd<<8)|x.c0;
-         m->add((stretch(StateMaps[0].p( dd,x.y))+1)>>1);
-        U32 c3=x.buf(3), c;
-        c=(x.words>>1)&63;
-        m->set(x.c0, 256);
-        m->set(order << 3U | x.bpos, 64);
-        m->set((x.bc4&3)*64+c+order*256, 256*8);
-        m->set(256*order + (x.w4&240) + (x.b3>>4), 256*8);
-        m->set((x.w4&255)+256*x.bpos, 256*8);
-        if (x.bpos){
-            c=x.c0<<(8-x.bpos); if (x.bpos==1)c+=x.b4/2;
-            c=(min(x.bpos,5))*256+(x.tt&63)+(c&192);
-        }
-        else c=(x.words&12)*16+(x.tt&63);
-        m->set(c, 1536);
-        c=x.bpos*256+((x.c0<<(8-x.bpos))&255);
-        c3 = (x.words<<x.bpos) & 255;
-        m->set(c+(c3>>x.bpos), 2048);
-        pr0=m->p(1,1);
-        int limit=0x3FF>>((x.blpos<0xFFF)*2);
-    int pr1, pr2, pr3;
-
-    pr  = Text.APMs[0].p(pr0, (x.c0<<8)|(x.Text.mask&0xF)|((x.Misses&0xF)<<4), x.y, limit);
-    pr1 = Text.APMs[1].p(pr0, x.c0^hash(x.bpos, x.Misses&3, x.buf(1), x.x5&0x80ff, x.Text.mask>>4)&0xFFFF, x.y, limit);
-    pr2 = Text.APMs[2].p(pr0, x.c0^hash( x.Match.byte, min(3, ilog2(x.Match.length+1)))&0xFFFF, x.y, limit);
-    pr3 = Text.APMs[3].p(pr0, x.c0^hash( x.c4&0xffff,   x.Text.firstLetter,x.wrtbytesize)&0xFFFF, x.y, limit);
-    pr0 = (pr0+pr1+pr2+pr3+2)>>2;
-
-    pr1 = Text.APM1s[0].p(pr0, x.c0^hash(x.Match.byte, min(3, ilog2(x.Match.length+1)), x.buf(1))&0xFFFF);
-    pr2 = Text.APM1s[1].p(pr, x.c0^hash( x.c4&0x00ffffff)&0xFFFF, 6);
-    pr3 = Text.APM1s[2].p(pr, x.c0^hash( x.c4&0xffffff00)&0xFFFF, 6);
-    pr = (pr+pr1+pr2+pr3+2)>>2;
-    pr = (pr+pr0)>>1;
-    sse.update();
-    pr = sse.p(pr);
-}
-
-// 1-bit image predicor
-class PredictorIMG1: public Predictors {
-  int pr;  // next prediction
-  Mixer *m;
-  eSSE sse;
-  const U8 activeModels[4] = { 
-   M_MATCH ,
-   M_MATCH1, 
-   M_IM1,
-   M_LSTM    };
-public:
-  int p()  const {/*assert(pr>=0 && pr<4096);*/ return pr;} 
-   ~PredictorIMG1(){ }
-
-PredictorIMG1(): pr(16384), sse(x) {
-   loadModels(activeModels,4);  
-   // add extra 
-   mixerInputs+=1;
-   mixerNets+=  256;
-   mixerNetsCount+=1;
-   sse.p(pr);
-   // create mixer
-   m=new Mixer(mixerInputs,  mixerNets,x, mixerNetsCount,0,3,2);
-}
-
-void update()  {
-  m->update();
-  m->add(256); 
-  int ismatch=models[M_MATCH]->p(*m);
-  models[M_MATCH1]->p(*m);
-  if (slow==true) models[M_LSTM]->p(*m);
-  m->set(ismatch,256);
-  models[M_IM1]->p(*m, x.finfo);
-  pr=m->p(); 
-  sse.update();
-    pr = sse.p(pr);
-}
-};
-
-// Audio predicor
-class PredictorAUDIO2: public Predictors {
-  int pr;
-  Mixer *m;
-  EAPM a;
-  eSSE sse;
-  const U8 activeModels[3] = {
-   M_RECORD ,
-   M_MATCH ,    
-   M_WAV  };
-  void setmixer();
-public:
-  int p()  const {/*assert(pr>=0 && pr<4096);*/ return pr;} 
-   ~PredictorAUDIO2(){  }
-
-PredictorAUDIO2(): pr(16384),a(x), sse(x) {
-   loadModels(activeModels,3);  
-   // add extra 
-   mixerInputs+=1;
-   mixerNets+=0;
-   mixerNetsCount+=0;
-   sse.p(pr);
-   // create mixer
-   m=new Mixer(mixerInputs,  mixerNets,x, mixerNetsCount);
-}
-
-void update()  {
-  m->update();
-  m->add(256);
-  models[M_MATCH]->p(*m);  
-  models[M_RECORD]->p(*m);
-  models[M_WAV]->p(*m,x.finfo);
-  pr=(32768-pr)/(32768/4096);
-  if(pr<1) pr=1;
-  if(pr>4095) pr=4095;
-  pr=a.p1(m->p(((x.finfo&2)==0),1),pr,7);
-  sse.update();
-    pr = sse.p(pr);
-}
-};
 //////////////////////////// Encoder ////////////////////////////
 
 // An Encoder does arithmetic encoding.  Methods:
